More work on the usability for CW.
Added byttons for typical QSO phrases.

Replaced the fractional resampler in the CW feedbach chain with a rational resampler that has a lot less latency.
Replaces the output autio switch with a matric multiplicato which eliminated the different latency between
the LSB and the USB chain, and therefore I also removed the delay blocks and the delay control widget.

Removed the delay block in the CW chain. It did not seem necessarey anymore.