Changed the Message Mangler to be able to apply a factor before adding the offset.
Addes the mode None to signify no transmission allowed.

Updated band plans to state None mode explicitly
Added automation to select the first legal bode specified in the band plan that is available when band selector is changed.

Added tx frequency to the debug values.

Reconfigured the waterfall display to have the same bandwidth as the sampling frequency. This makes the WF i bit more understandable.