Changed the message gate to make it possible to repeat outgoing messages in order to work around a bug in the QT GU Digital Number control. To make it actually work, it was necessary to split each message with 0.1 seconds. This might actually be shorter but 0.1 seconds is good for now.
Added a chooser for the RX sampling frequency, it will be removed because the sampling fq cannot be changed on the fly.

Moved initialization of the band selector box to the after_start snippet because otherwise startup crashed.

Lowered the tx sampling rate to 3 MHz,

Added a 1.3 kHz Rx bandwidth option specially designed for MSK144 operation.

The old fbq_xcvr GRC now generates its own file.